单独提出了table对象到独立的文件
